Change duration on SQL(Bulk)Response from int to float
In order to have more precision without changing from `ms` to something
else. (Otherwise the duration in Crash would be broken)

+++ sql/src/main/java/io/crate/action/sql/SQLBulkResponse.java
@@ -40,7 +40,7 @@ public class SQLBulkResponse extends SQLBaseResponse {
 
     public SQLBulkResponse(String[] outputNames,
                            Result[] results,
-                           int duration,
+                           float duration,
                            DataType[] colTypes,
                            boolean includeTypes) {
         super(outputNames, colTypes, includeTypes, duration);
